kvserver: move addWriteBatch
The next commit will remove its dependency on `replicaAppBatch`.

Epic: CRDB-220
Release note: None
